WeatherClimateAnalytics/src/requirements.txt
# requirements.txt
streamlit
snowflake-connector-python
matplotlib
pandas
plotly
numpy
